for #42: Create abstract base class from reusable RRD bits.
Currently, the RoundRobinDispatcher implementation is the only Dispatcher implementation that's available. However, it is desirable to add others.

To facilitate this, the reusable parts of RoundRobinDispatcher are pulled out in a new, abstract superclass. New Dispatcher implementations can choose to inherit from that.
